Compiled beta test version!
This is a compiled version of the code. Please don't judge it strictly, as it's beta and not fully functional. If possible, I'll fix the code. If you need it, just write your code in code.pialn. Use Python, but it's important that you won't be able to use your libraries unless you recompile the code. Maybe I'll release a version for this where you can add libraries, but with such an online environment, it might take some time.